have binary path resolution on disk working (very naive tho)

+++ src/kernel/src/drivers/dt.c
@@ -69,7 +69,7 @@ void device_tree_init(){
     u64 tree = get_device_tree_start();
     device_trie = trie_init();
     trie_add(device_trie, "/dev/fb0", 1);
-    trie_add(device_trie, "/dev/ttyS1", &keyboard_file);
+    trie_add(device_trie, "/dev/ttyS1", (u64) &keyboard_file);
 
     if(FALSE){
         fdt_header* header = (fdt_header*) tree;
